+++ app/src/main/java/com/cajetan/youtubeplus/data/VideoData.java
@@ -1,0 +1,14 @@
+package com.cajetan.youtubeplus.data;
+
+import android.arch.persistence.room.ColumnInfo;
+import android.arch.persistence.room.Entity;
+import android.arch.persistence.room.PrimaryKey;
+
+@Entity
+public class VideoData {
+    @PrimaryKey
+    public int vIndex;
+
+    @ColumnInfo(name = "video_id")
+    public String videoId;
+}